[TimePicker] Sort out tests
