+++ packages/material-ui-lab/src/TimePicker/TimePicker.spec.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import moment from 'moment';
-import { TimePicker, ClockPicker } from '@material-ui/lab';
+import TimePicker from '@material-ui/lab/TimePicker';
 
 <TimePicker
   value={moment()}
@@ -15,6 +15,3 @@ import { TimePicker, ClockPicker } from '@material-ui/lab';
   onChange={(date) => date?.set({ second: 0 })}
   renderInput={() => <input />}
 />;
-
-// External components are generic as well
-<ClockPicker<Date> view="hours" date={null} onChange={(date) => date?.getDate()} />;
